pyproject.toml
[project]
name = "pretraining"
version = "0.1.0"
description = "DINOv2 → YOLO knowledge distillation for pretraining detection backbones"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["deep-learning", "distillation", "yolo", "dinov2", "self-supervised", "pytorch-lightning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "hydra-colorlog>=1.2.0",
    "pillow>=9.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
yolo = [
    "ultralytics>=8.0.0",
]
wandb = [
    "wandb>=0.15.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
all = [
    "pretraining[yolo,wandb,dev]",
]

[project.scripts]
train = "src.train:main"
export = "src.export:main"

[project.urls]
Homepage = "https://github.com/yourusername/pretraining"
Documentation = "https://github.com/yourusername/pretraining#readme"
Repository = "https://github.com/yourusername/pretraining"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
